[fix/search-prefix/refactor2] move PersistentValue to SearchUtils And delete SearchState.js
